migrate proxy unit tests to pytest
